hardware/requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.6.0
Pillow==6.2.1
pylint==2.4.4
python-ev3dev2==2.0.0b5
regex==2019.11.1
requests==2.22.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
